requirements.txt
contextvars
selenium
pytest
allure-pytest
pyyaml
pytest-rerunfailures
